Rename file name to func_utils.py
