+++ pymc/func_utils.py
